+++ components/molecules/Card.stories.js
@@ -1,0 +1,30 @@
+import React from "react";
+import Card from "./Card";
+
+export default {
+  title: "Components/Molecules/Card",
+  component: Card,
+};
+
+const Template = (args) => <Card {...args} />;
+
+export const Primary = Template.bind({});
+export const Experiment = Template.bind({});
+
+Primary.args = {
+  title: "Title",
+  description: "Description",
+  imgSrc: "/placeholderImg",
+  imgAlt: "placeholderAlt",
+};
+
+Experiment.args = {
+  isExperiment: true,
+  title: "Title",
+  tag: "experiment_tag",
+  tagLabel: "Experiment tag",
+  description: "Description",
+  href: "/some/link",
+  imgSrc: "/placeholderImg",
+  imgAlt: "placeholderAlt",
+};