+++ CHANGELOG.md
@@ -31,6 +31,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Added the `stripFeedback` helper function which strips PII from both the feedback and report a problem forms
 - Added error label in report a problem widget
 - Added environment variable for our project version to our Next.js config for metadata purposes
+- Added image in Experiment card
 
 ## Changed
 
@@ -59,6 +60,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Updated `/notsupported`, `/404` and `/500` to use data from AEM
 - Updated `/signup` page to use get data from AEM
 - Updated `/notsupported`, `/404`, `/500`, and `/error` to use data from AEM
+- Updated Experiment component to a generic card component
 
 ## Fixed